Fix: changed type of size and offset to off_t
As the parameters for map read from JavaScript where explicitely set to
int only 2GB could be used for either size or offset even on a 64bit system.
The type off_t automatically maps to a int32_t on 32 bit systems and
int64_t on 64 bit. The original size_t maps to unsigned which
would have been more appropriate but NAN does not contain an
unsigned type for 64bit.